Fix usability around CLI
The cli is super straightforward, with just a single argument. But, that
doesn't mean people won't forget the config file or specify extra args
by mistake. This commit adds a --help/-h flag which will print the usage
information. It also will handle the case when an incorrect number of
arguments are supplied and refer the user to --help.